Fit risk ratio and risk difference models

riskratio and riskdiff provide a flexible interface to fitting risk ratio and risk difference models.

In cohort studies with a binary outcome, risk ratios and risk differences are typically more appropriate to report than odds ratios from logistic regression, yet such models have historically been difficult to implement in standard software.

The risks package selects an efficient way to fit risk ratio or risk difference models successfully, which will converge whenever logistic models converge. Optionally, a specific approach to model fitting can also be requested. Implemented are Poisson models with robust covariance, binomial models, binomial models aided in convergence by starting values obtained through Poisson models, binomial models fitted via combinatorial expectation maximization (optionally also with Poisson starting values), and estimates obtained via marginal standardization after logistic regression.

Adjusting for covariates (e.g., confounders) in the model specification (formula =) is possible.

riskratio(
  formula,
  data,
  approach = c("auto", "all", "robpoisson", "glm", "glm_start", "glm_cem",
    "glm_cem_start", "margstd", "logistic"),
  variable = NULL,
  at = NULL,
  ...
)

riskdiff(
  formula,
  data,
  approach = c("auto", "all", "robpoisson", "glm", "glm_start", "glm_cem",
    "glm_cem_start", "margstd", "logistic"),
  variable = NULL,
  at = NULL,
  ...
)

Arguments

formula
A formula object of the form 'response ~ predictors'.

data
A tibble or data.frame object.

approach
Optional: Method for model fitting.

"auto" (default) is recommended; it selects the most efficient approach that converges and ensures that predicted probabilities are within range (< 1; see Details).

"all" will attempt to fit the model via all implemented approaches to allow for comparisons.

The other options allow for directly selecting a fitting approach, some of which may not converge or yield out-of-range predicted probabilities. See full documentation (currently at https://stopsack.github.io/risks) for details.

"glm" Binomial model.

"glm_start" Binomial model with starting values from Poisson model.

"robpoisson" Poisson model with robust covariance.

"glm_cem" Binomial model fitted with combinatorial expectation maximization.

"glm_cem_start" As glm_cem, with Poisson starting values.

"margstd" Marginal standardization after logistic model.

"logistic" For comparison only: the logistic model. Only available in riskratio().

variable
Optional: exposure variable to use for marginal standardization. If variable is not provided and marginal standardization is attempted, then the first binary or categorical variable in the model is used as the exposure. Levels are determined automatically for variables types logical, character, factor, and numeric (only if no more than 2 levels for the latter); otherwise levels must be supplied via at =.

at
Optional: Levels of exposure variable variable for marginal standardization. at = determines the levels at which contrasts of the exposure are to be assessed. The level listed first is used as the reference. Levels must exist in the data for character, factor or ordered factor variables. For numeric variables, levels that do not exist in the data can be interpolations or extrapolations; if levels exceed the extremes of the data (extrapolation), a warning will be displayed.

...
Optional: Further arguments passed to fitting functions (glm, logbin, or addreg).

Value

Fitted model. This object can be passed on to post-processing functions:

summary.risks: an overview of results (risks-specific S3 methods: summary.robpoisson, summary.margstd).

tidy.risks: a tibble of coefficients and confidence intervals.

Standard post-processing functions can also be used:

coef: a vector of coefficients.

confint: a matrix of confidence intervals (risks-specific S3 methods: confint.robpoisson, confint.margstd).

predict.glm(type = "response"): fitted values (predictions).

residuals: residuals.

If model fitting using all possible approaches was requested via approach = "all", then their results can be retrieved from the list all_models in the returned object (e.g., fit$all_models[[1]], fit$all_models[[2]], etc.).

Functions

riskratio: Fit risk ratio models

riskdiff: Fit risk difference models

References

Wacholder S. Binomial regression in GLIM: Estimating risk ratios and risk differences. Am J Epidemiol 1986;123:174-184. (Binomial regression models; approach = "glm")

Spiegelman D, Hertzmark E. Easy SAS Calculations for Risk or Prevalence Ratios and Differences. Am J Epidemiol 2005;162:199-200. (Binomial models fitted used starting values from Poisson models; approach = "glm_start")

Zou G. A modified Poisson regression approach to prospective studies with binary data. Am J Epidemiol 2004;159:702-706. (Poisson model with robust/sandwich standard errors; approach = "robpoisson")

Donoghoe MW, Marschner IC. logbin: An R Package for Relative Risk Regression Using the Log-Binomial Model. J Stat Softw 2018;86(9). (Log-binomial models fitted via combinatorial expectation maximization; riskratio(approach = "glm_cem")

Donoghoe MW, Marschner IC. Stable computational methods for additive binomial models with application to adjusted risk differences. Comput Stat Data Anal 2014;80:184-96. (Additive binomial models fitted via combinatorial expectation maximization; riskdiff(approach = "glm_cem"))

Localio AR, Margolis DJ, Berlin JA. Relative risks and confidence intervals were easily computed indirectly from multivariable logistic regression. J Clin Epidemiol 2007;60(9):874-82. (Marginal standardization after fitting a logistic model; approach = "margstd")

Examples

# Newman SC. Biostatistical methods in epidemiology. New York, NY: Wiley, 2001, table 5.3
library(tibble)  # used to set up example data
dat <- tibble(
  death    = c(rep(1, 54), rep(0, 138)),
  stage    = c(rep("Stage I", 7),  rep("Stage II", 26), rep("Stage III", 21),
               rep("Stage I", 60), rep("Stage II", 70), rep("Stage III", 8)),
  receptor = c(rep("Low", 2),  rep("High", 5),  rep("Low", 9),  rep("High", 17),
               rep("Low", 12), rep("High", 9),  rep("Low", 10), rep("High", 50),
               rep("Low", 13), rep("High", 57), rep("Low", 2),  rep("High", 6)))

# Risk ratio model
fit_rr <- riskratio(formula = death ~ stage + receptor, data = dat)
fit_rr

#> 
#> Risk ratio model
#> Call:  stats::glm(formula = death ~ stage + receptor, family = binomial(link = "log"), 
#>     data = data, start = "(from Poisson model)")
#> 
#> Coefficients:
#>    (Intercept)   stageStage II  stageStage III     receptorLow  
#>        -2.3521          0.9314          1.7695          0.4436  
#> 
#> Degrees of Freedom: 191 Total (i.e. Null);  188 Residual
#> Null Deviance:	    228.1 
#> Residual Deviance: 185.9 	AIC: 193.9

summary(fit_rr)

#> 
#> Risk ratio model, fitted as binomial model with starting values from Poisson model (glm_start).
#> Call:
#> stats::glm(formula = death ~ stage + receptor, family = binomial(link = "log"), 
#>     data = data, start = "(from Poisson model)")
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -2.0209  -0.7436  -0.4472   0.5272   2.1689  
#> 
#> Coefficients:
#>                Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)     -2.3521     0.3602  -6.531 6.55e-11 ***
#> stageStage II    0.9314     0.3930   2.370   0.0178 *  
#> stageStage III   1.7695     0.3851   4.595 4.33e-06 ***
#> receptorLow      0.4436     0.1968   2.254   0.0242 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 228.15  on 191  degrees of freedom
#> Residual deviance: 185.85  on 188  degrees of freedom
#> AIC: 193.85
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#> Confidence intervals for coefficients: (normality-based)
#>                      2.5 %     97.5 %
#> (Intercept)    -3.05805977 -1.6461981
#> stageStage II   0.16107688  1.7018004
#> stageStage III  1.01475475  2.5242885
#> receptorLow     0.05781282  0.8294095

# Risk difference model
fit_rd <- riskdiff(formula = death ~ stage + receptor, data = dat)
fit_rd

#> 
#> Risk difference model
#> Call:  stats::glm(formula = death ~ stage + receptor, family = binomial(link = "identity"), 
#>     data = data, start = "(no starting values)")
#> 
#> Coefficients:
#>    (Intercept)   stageStage II  stageStage III     receptorLow  
#>        0.08381         0.14921         0.57227         0.16131  
#> 
#> Degrees of Freedom: 191 Total (i.e. Null);  188 Residual
#> Null Deviance:	    228.1 
#> Residual Deviance: 186.4 	AIC: 194.4

summary(fit_rd)

#> 
#> Risk difference model, fitted as binomial model (glm).
#> Call:
#> stats::glm(formula = death ~ stage + receptor, family = binomial(link = "identity"), 
#>     data = data, start = "(no starting values)")
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.8441  -0.7284  -0.4184   0.6350   2.2268  
#> 
#> Coefficients:
#>                Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)     0.08381    0.03633   2.307  0.02107 *  
#> stageStage II   0.14921    0.05755   2.593  0.00953 ** 
#> stageStage III  0.57227    0.09472   6.042 1.52e-09 ***
#> receptorLow     0.16131    0.07587   2.126  0.03349 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 228.15  on 191  degrees of freedom
#> Residual deviance: 186.38  on 188  degrees of freedom
#> AIC: 194.38
#> 
#> Number of Fisher Scoring iterations: 5
#> 
#> Confidence intervals for coefficients: (normality-based)
#>                     2.5 %    97.5 %
#> (Intercept)    0.01260046 0.1550111
#> stageStage II  0.03640945 0.2620070
#> stageStage III 0.38663167 0.7579168
#> receptorLow    0.01261165 0.3100162