Tidy model summaries for risks models

Obtain a tibble (data frame) with parameters, coefficients, standard errors, confidence limits, and p-values. A column with the type of model fitted is added.

# S3 method for risks
tidy(
  x,
  conf.int = TRUE,
  conf.level = 0.95,
  bootrepeats = 200,
  exponentiate = FALSE,
  default = TRUE,
  ...
)

Arguments

x
Model

conf.int
Show confidence intervals?

conf.level
Optional: Confidence level. Defaults to 0.95.

bootrepeats
Optional: Number of bootstrap repeats. Applicable to models fitted via marginal standardization (approach = "margstd"). Defaults to 200, a very low number of repeats. It is strongly recommended to increase repeats to >1000.

exponentiate
Exponentiate coefficients and confidence limits? Defaults to FALSE. Setting exponentiate = TRUE is useful for relative risk models (log links).

default
Use normality-based confidence intervals? Defaults to TRUE. With default = FALSE, profile likelihood-based confidence intervals can be calculated for binomial models.

...
Passed on

Value

tibble

Details

If multiple types of models are fitted, tidy() can be used to parameters for all models at once, in one tibble. The last column of the tibble includes the name of the model. See examples.

Examples

# Define example data
library(tibble)  # for tibble() function
library(broom)   # required for tidy.risks()
dat <- tibble(
  death    = c(rep(1, 54), rep(0, 138)),
  stage    = c(rep("Stage I", 7),  rep("Stage II", 26), rep("Stage III", 21),
               rep("Stage I", 60), rep("Stage II", 70), rep("Stage III", 8)),
  receptor = c(rep("Low", 2),  rep("High", 5),  rep("Low", 9),  rep("High", 17),
               rep("Low", 12), rep("High", 9),  rep("Low", 10), rep("High", 50),
               rep("Low", 13), rep("High", 57), rep("Low", 2),  rep("High", 6)))
# Fit and tidy the model
fit_rr <- riskratio(formula = death ~ stage + receptor, data = dat)
tidy(fit_rr)

#> # A tibble: 4 x 8
#>   term         estimate std.error statistic  p.value conf.low conf.high model   
#>   <chr>           <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl> <chr>   
#> 1 (Intercept)    -2.35      0.360     -6.53 6.55e-11  -3.06      -1.65  glm_sta…
#> 2 stageStage …    0.931     0.393      2.37 1.78e- 2   0.161      1.70  glm_sta…
#> 3 stageStage …    1.77      0.385      4.60 4.33e- 6   1.01       2.52  glm_sta…
#> 4 receptorLow     0.444     0.197      2.25 2.42e- 2   0.0578     0.829 glm_sta…

# Marginal standardization,
# increase number of bootstrap repeats:
fit_rr <- riskratio(formula = death ~ stage + receptor, data = dat, approach = "margstd")
tidy(fit_rr, bootrepeats = 1000)

#> # A tibble: 3 x 8
#>   term           estimate std.error statistic p.value conf.low conf.high model  
#>   <chr>             <dbl>     <dbl>     <dbl>   <dbl>    <dbl>     <dbl> <chr>  
#> 1 stageStage I      0          0      NaN     NaN       NA         NA    margstd
#> 2 stageStage II     0.899      1.27     0.709   0.478    0.169      1.83 margstd
#> 3 stageStage III    1.81       1.28     1.41    0.158    1.11       2.77 margstd

# Multiple types of models fitted:
fit_rr <- riskratio(formula = death ~ stage + receptor, data = dat, approach = "all")

#> Loading required package: doParallel

#> Loading required package: foreach

#> Loading required package: iterators

#> Loading required package: parallel

#> Loading required package: numDeriv

#> Loading required package: quantreg

#> Loading required package: SparseM

#> 
#> Attaching package: ‘SparseM’

#> The following object is masked from ‘package:base’:
#> 
#>     backsolve

#> 
#> Attaching package: ‘turboEM’

#> The following objects are masked from ‘package:numDeriv’:
#> 
#>     grad, hessian

tidy(fit_rr)

#> # A tibble: 23 x 8
#>    term        estimate std.error statistic  p.value conf.low conf.high model   
#>    <chr>          <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl> <chr>   
#>  1 (Intercept)   -2.37      0.385     -6.14 8.34e-10  -3.07      -1.66  robpois…
#>  2 stageStage…    0.925     0.426      2.17 3.01e- 2   0.154      1.70  robpois…
#>  3 stageStage…    1.78      0.447      3.97 7.10e- 5   1.02       2.53  robpois…
#>  4 receptorLow    0.489     0.288      1.70 8.89e- 2   0.0718     0.906 robpois…
#>  5 (Intercept)   -2.35      0.360     -6.53 6.55e-11  -3.06      -1.65  glm_sta…
#>  6 stageStage…    0.931     0.393      2.37 1.78e- 2   0.161      1.70  glm_sta…
#>  7 stageStage…    1.77      0.385      4.60 4.33e- 6   1.01       2.52  glm_sta…
#>  8 receptorLow    0.444     0.197      2.25 2.42e- 2   0.0578     0.829 glm_sta…
#>  9 (Intercept)   -2.35      0.360     -6.53 6.55e-11  -3.06      -1.65  logbin  
#> 10 stageStage…    0.931     0.393      2.37 1.78e- 2   0.161      1.70  logbin  
#> # … with 13 more rows